backend/requirements-prod.txt
fastapi==0.115.5
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
httpx==0.28.1
python-dotenv==1.2.1
